Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,6 +4,6 @@
 int main() {
     int r;
     scanf("%d",&r);
-    printf("Area: %d",(float)(PI*(r**2));
+    printf("Area: %.2f",(PI*(r**2));
     return 0;
 }
